refactor(subscriptions): split SubscriptionCard into focused sub-components Convert SubscriptionCard.tsx into a SubscriptionCard/ folder with three dedicated sub-components, each responsible for a single visual section:   SubscriptionCard/     index.tsx                   orchestrator ~30 lines: useSWR + compose     SubscriptionCardHeader.tsx  Avatar, name, and 3 action icon-buttons     SubscriptionNextPayment.tsx next-payment row: date+chip or italic note                                 getDaysChipColor/getDaysLabel kept private     SubscriptionPaymentList.tsx Ultimos pagos section: list + unlink btns Public API (Props) and all existing tests are unchanged. SubscriptionCard.test.tsx resolves the new index.tsx automatically. Also removes the accidentally committed packages/api/globalConfig.json.
